requirements.txt
flask
ultralytics
pillow
torch
torchvision
requests
